+++ apps/block_scout_web/lib/block_scout_web/views/api/v2/token_view.ex
@@ -1,0 +1,13 @@
+defmodule BlockScoutWeb.API.V2.TokenView do
+  def render("token.json", %{token: token}) do
+    %{
+      "address" => token.contract_address_hash,
+      "symbol" => token.symbol,
+      "name" => token.name,
+      "decimals" => token.decimals,
+      "type" => token.type,
+      "holders" => to_string(token.holder_count),
+      "exchange_rate" => token.usd_value && to_string(token.usd_value)
+    }
+  end
+end